python/gpio/requirements.txt
RPi.GPIO
python3-smbus
